requirements.txt
fastapi
uvicorn
httpx
pandas
numpy
scikit-learn
python-dotenv